requirements.txt
awsebcli==3.10.4
blessed==1.14.2
boto3==1.4.4
botocore==1.5.85
cement==2.8.2
colorama==0.3.7
Django==1.10.7
django-cleanup==1.0.1
django-comments-xtd==2.0.5
django-contrib-comments==1.8.0
django-el-pagination==3.1.0
django-ipware==1.1.6
django-storages==1.6.3
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.3
olefile==0.44
pathspec==0.5.0
Pillow==4.2.1
psycopg2==2.7.1
python-dateutil==2.6.1
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.10
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.44.0
